Move expanded view components into new directory
