+++ client/components/MediaContainer.jsx
@@ -4,13 +4,13 @@ import axios from 'axios';
 import MainImage from './MainImage.jsx';
 import AdditionalMedia from './AdditionalMedia.jsx';
 import Caption from './Caption.jsx';
-import ExpandedView from './ExpandedView.jsx';
+import ExpandedView from './modal/ExpandedView.jsx';
 
 class MediaContainer extends Component {
   constructor() {
     super();
     this.state = {
-      id: 47,
+      id: 46,
       name: '',
       images: [],
       main: '',
