+++ client/components/modal/ExpandedImageContent.jsx
