+++ client/components/modal/ExpandedThumbnail.jsx
