+++ client/components/modal/ExpandedThumbnails.jsx
